Move to single package repo structure
