+++ docs/content/1.getting-started/3.contributing.md
@@ -30,13 +30,13 @@ PRs are very welcome. Please be mindful of breaking changes. If you're thinking 
 
 This allows you to mess around with the component in isolation to quickly prototype and iterate on your change.
 
-`pnpm -F vue-split-panel playground`
+`pnpm playground`
 
 ## Writing docs
 
 The documentation is a [Docus](https://docus.dev)-based Nuxt application. You can run it in development with:
 
-`pnpm -F docs dev`
+`pnpm docs:dev`
 
 Please make sure to update the docs if you're implementing new features, or changing the default behavior.
 
@@ -44,4 +44,4 @@ Please make sure to update the docs if you're implementing new features, or chan
 
 `vue-split-panel` relies on vitest for it's tests. You can run them with
 
-`pnpm -F vue-split-panel test`
+`pnpm test`
