--- docs/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "docs",
-	"private": true,
-	"scripts": {
-		"dev": "nuxt dev",
-		"build": "nuxt build"
-	},
-	"dependencies": {
-		"@directus/vue-split-panel": "workspace:*",
-		"@nuxt/kit": "4.3.0",
-		"@nuxt/ui-pro": "3.3.7",
-		"better-sqlite3": "12.6.2",
-		"docus": "5.4.4",
-		"nuxt": "4.3.0",
-		"tailwindcss": "4.1.18",
-		"vue": "3.5.27"
-	},
-	"devDependencies": {
-		"@nuxtjs/mdc": "0.20.0",
-		"unist-util-visit": "5.1.0"
-	}
-}
