--- packages/vue-split-panel/package.json
@@ -1,65 +1,0 @@
-{
-	"name": "@directus/vue-split-panel",
-	"version": "0.8.12",
-	"description": "Split panel component for Vue based on WebAwesome Split Panel",
-	"keywords": [
-		"components",
-		"panel",
-		"splitter",
-		"ui",
-		"vue"
-	],
-	"homepage": "https://github.com/directus/vue-split-panel",
-	"bugs": {
-		"url": "https://github.com/directus/vue-split-panel/issues"
-	},
-	"author": "Rijk van Zanten <rijkvanzanten@me.com>",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/directus/vue-split-panel.git"
-	},
-	"files": [
-		"dist"
-	],
-	"type": "module",
-	"sideEffects": false,
-	"module": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		".": {
-			"import": "./dist/index.js"
-		},
-		"./package.json": "./package.json",
-		"./index.css": "./dist/style.css"
-	},
-	"scripts": {
-		"build": "vp pack",
-		"dev": "vp pack --watch",
-		"playground": "vp dev",
-		"test": "vp test --run",
-		"test:coverage": "vp test --run --coverage",
-		"release": "bumpp && pnpm publish",
-		"prepublishOnly": "pnpm run build",
-		"lint": "vp lint --type-aware --type-check",
-		"lint:fix": "vp lint --fix"
-	},
-	"dependencies": {
-		"@vueuse/core": "14.3.0"
-	},
-	"devDependencies": {
-		"@tsdown/css": "0.22.0",
-		"@types/node": "25.9.1",
-		"@vitejs/plugin-vue": "6.0.7",
-		"@vitest/coverage-v8": "4.1.7",
-		"@vue/test-utils": "2.4.10",
-		"bumpp": "11.1.0",
-		"happy-dom": "20.9.0",
-		"typescript": "6.0.3",
-		"vite-plus": "catalog:",
-		"vue": "3.5.35",
-		"vue-tsc": "3.2.9"
-	},
-	"peerDependencies": {
-		"vue": "^3.5.0"
-	}
-}
