+++ src/utils/percentage-to-pixels.test.ts
