+++ src/utils/pixels-to-percentage.test.ts
