+++ src/utils/pixels-to-percentage.ts
