+++ tests/collapse.test.ts
